refactor(client): extract TwoFactorCard into standalone component
